requirements.txt
selenium==4.27.1
pytest==8.3.4
pytest-html==4.1.1
webdriver-manager==4.0.2
allure-pytest==2.15.0
requests==2.32.5